move shared code to merger
