--- modules/chameleon/src/main/java/org/clueminer/chameleon/MergingStrategy.java
@@ -1,10 +1,0 @@
-package org.clueminer.chameleon;
-
-/**
- *
- * @author Tomas Bruna
- */
-public enum MergingStrategy {
-
-    MULTIPLE, PAIR
-}
